backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.1.0
sqlalchemy>=2.0.0
alembic>=1.13.0
flask-migrate>=4.0.0
requests>=2.31.0
PyJWT>=2.8.0
bcrypt>=4.1.0
pytest>=8.0.0
pytest-flask>=1.3.0
